Move dropdown-setting to same location as other core callbacks
This probably wants a further refactor with respect to where functions
are defined, but is useful for now to avoid circular imports; and it
makes sense for them to be in one place
